+++ app/src/main/java/com/mcode/mercadolibre/models/api/ReferencePrices.kt
@@ -7,7 +7,7 @@ data class ReferencePrices (
 	@SerializedName("id") val id : Int,
 	@SerializedName("type") val type : String,
 	@SerializedName("conditions") val conditions : Conditions,
-	@SerializedName("amount") val amount : Int,
+	@SerializedName("amount") val amount : Float,
 	@SerializedName("currency_id") val currencyId : String,
 	@SerializedName("exchange_rate_context") val exchangeRateContext : String,
 	@SerializedName("tags") val tags : List<String>,
